Fifth Commit,
Added ability to allow user sign-in only after verifying the email entered during sign up process. Also if the account is deleted from the firebase console, the now "un-authorized" user will be kicked out in an hour when the token refreshes automatically or when the app is launched twice next times as the token will be reloaded forcefully everytime user launches the app, so in case if the user launches app i.e. first launch after deletion of the account and the token could not be refreshed, next time the user launches the app i.e. second launch the user will be kicked out instantaneously without waiting for the 1 hour time limit.

Added HomeActivity (blank for now), if user is signed-in, HomeActivity is launched almost instantaneously, and when user is logged out, the sign-in activity kicks in automatically as it is the launcher activity.

Signed-off-by: Jai Kherajani <jai.kherajani19@gmail.com>